requirements.txt
discord.py==2.3.2
undetected-chromedriver==3.5.5
